fix(mobile): 주문 상태 null 파싱 크래시 + 반려수량 소수 박스 절단 방지
주문상세/목록 응답의 서버-클라 계약 불일치 2건을 교정.

1) 주문 상태 non-null 캐스팅 크래시
   backend 는 orderRequestStatus/orderRequestStatusName 을 SF nillable=true
   정합으로 String? 로 내려주는데(마이그레이션 SF NULL row 보존), mobile 모델/
   엔티티가 `as String` non-null 캐스팅이라 NULL row 상세/목록 진입 시 파싱 예외.
   - OrderRequestDetailModel / OrderRequestModel / OrderDetail / OrderRequest 의
     두 status 필드를 String? 로 전환 + fromJson `as String?`.
   - OrderStatusCode.color / OrderRequestStatusBadge 를 nullable 수용(null → 회색,
     표시명 null → 빈 라벨). == 비교는 어느 상태 코드와도 불일치로 안전.

2) 반려수량 int 절단
   backend RejectedItemResponse.orderQuantityBoxes 는 BigDecimal 인데 mobile 이
   `as num).toInt()` 로 받아 소수 박스에서 절단/파싱 리스크. OrderedItem.
   totalQuantityBoxes 와 동일하게 double 로 전환 + NumberFormat('#,##0.##') 표시.

+++ mobile/test/helpers/fake_order_request_repository.dart
@@ -504,7 +504,7 @@ class FakeOrderRequestRepository implements OrderRequestRepository {
         .map((p) => RejectedItem(
               productCode: p.$1,
               productName: p.$2,
-              orderQuantityBoxes: p.$3,
+              orderQuantityBoxes: p.$3.toDouble(),
               rejectionReason: p.$4,
             ))
         .toList();
